+++ user-service/prototype/src/main/java/jp/co/itmeister/userservice/userservice/entity/PostEntity.java
@@ -56,7 +56,7 @@ public class PostEntity {
     private ZonedDateTime snapTime;
 
     @Column(name = "angle")
-    private Integer angle;
+    private float angle;
 
     @Column(name = "iso", length = 32)
     private String iso;
@@ -157,11 +157,11 @@ public class PostEntity {
         this.snapTime = snapTime;
     }
 
-    public Integer getAngle() {
+    public float getAngle() {
         return angle;
     }
 
-    public void setAngle(Integer angle) {
+    public void setAngle(float angle) {
         this.angle = angle;
     }
